requirements.txt
asgiref
Django
djangorestframework
Pillow
psycopg2-binary
pytz
sqlparse
typing_extensions
